ref: Moved httpget to misc
