+++ Environment/Libraries/Misc.cpp
@@ -4,10 +4,12 @@
 
 #include "Misc.hpp"
 
+#include <HttpStatus.hpp>
 #include <lz4.h>
 
 #include "Communication.hpp"
 #include "Scheduler.hpp"
+#include "cpr/api.h"
 
 namespace RbxStu {
     namespace Misc {
@@ -168,6 +170,42 @@ namespace RbxStu {
 
             return 0;
         }
+
+        int httpget(lua_State *L) {
+            const std::string url = luaL_checkstring(L, 1);
+
+            if (url.find("http://") == std::string::npos && url.find("https://") == std::string::npos)
+                luaL_argerror(L, 1, "Invalid protocol (expected 'http://' or 'https://')");
+
+            const auto scheduler = Scheduler::GetSingleton();
+
+            scheduler->ScheduleJob(SchedulerJob(
+                    L, [url](lua_State *L, std::shared_future<std::function<int(lua_State *)>> *callbackToExecute) {
+                        const auto response = cpr::Get(cpr::Url{url}, cpr::Header{{"User-Agent", "Roblox/WinInet"}});
+
+                        auto output = std::string("");
+
+                        if (HttpStatus::IsError(response.status_code)) {
+                            output = std::format("HttpGet failed\nResponse {} - {}. {}",
+                                                 std::to_string(response.status_code),
+                                                 HttpStatus::ReasonPhrase(response.status_code),
+                                                 std::string(response.error.message));
+                        } else {
+                            output = response.text;
+                        }
+
+                        *callbackToExecute =
+                                std::async(std::launch::async, [output]() -> std::function<int(lua_State *)> {
+                                    return [output](lua_State *L) -> int {
+                                        lua_pushlstring(L, output.c_str(), output.size());
+                                        return 1;
+                                    };
+                                });
+                    }));
+
+            L->ci->flags |= 1;
+            return lua_yield(L, 1);
+        }
     } // namespace Misc
 } // namespace RbxStu
 
@@ -184,6 +222,8 @@ luaL_Reg *Misc::GetLibraryFunctions() {
                               {"setclipboard", RbxStu::Misc::setclipboard},
                               {"getclipboard", RbxStu::Misc::getclipboard},
                               {"emptyclipboard", RbxStu::Misc::emptyclipboard},
+
+                              {"httpget", RbxStu::Misc::httpget},
 
                               {nullptr, nullptr}};
